[C] Move clock definitions to separate compilation unit.
