+++ aeron-driver/src/main/c/CMakeLists.txt
@@ -139,6 +139,7 @@ SET(SOURCE
     util/aeron_parse_util.c
     util/aeron_properties_util.c
     util/aeron_http_util.c
+    util/aeron_clock.c
     aeron_driver_context.c
     aeron_alloc.c
     aeron_driver.c
@@ -191,6 +192,7 @@ SET(HEADERS
     util/aeron_parse_util.h
     util/aeron_properties_util.h
     util/aeron_http_util.h
+    util/aeron_clock.h
     concurrent/aeron_thread.h
     concurrent/aeron_atomic.h
     concurrent/aeron_atomic64_gcc_x86_64.h
