+++ aeron-driver/src/main/c/aeron_driver.c
@@ -113,51 +113,6 @@ void aeron_log_func_none(const char *str)
 {
 }
 
-int64_t aeron_nano_clock()
-{
-    struct timespec ts;
-#if defined(__CYGWIN__) || defined(__linux__)
-    if (clock_gettime(CLOCK_MONOTONIC, &ts) < 0)
-    {
-        return -1;
-    }
-#elif defined(AERON_COMPILER_MSVC)
-    if (aeron_clock_gettime_monotonic(&ts) < 0)
-    {
-        return -1;
-    }
-#else
-    if (clock_gettime(CLOCK_MONOTONIC_RAW, &ts) < 0)
-    {
-        return -1;
-    }
-#endif
-
-    return (ts.tv_sec * 1000000000) + ts.tv_nsec;
-}
-
-int64_t aeron_epoch_clock()
-{
-    struct timespec ts;
-#if defined(AERON_COMPILER_MSVC)
-    if (aeron_clock_gettime_realtime(&ts) < 0)
-    {
-        return -1;
-    }
-#else
-#if defined(CLOCK_REALTIME_COARSE)
-    if (clock_gettime(CLOCK_REALTIME_COARSE, &ts) < 0)
-#else
-    if (clock_gettime(CLOCK_REALTIME, &ts) < 0)
-#endif
-    {
-        return -1;
-    }
-#endif
-
-    return (ts.tv_sec * 1000) + (ts.tv_nsec / 1000000);
-}
-
 extern int aeron_number_of_trailing_zeroes(int32_t value);
 extern int aeron_number_of_trailing_zeroes_u64(uint64_t value);
 extern int aeron_number_of_leading_zeroes(int32_t value);
