+++ aeron-driver/src/main/c/aeronmd.h
@@ -739,25 +739,6 @@ int aeron_driver_close(aeron_driver_t *driver);
 int aeron_delete_directory(const char *dirname);
 
 /**
- * Clock function used by aeron.
- */
-typedef int64_t (*aeron_clock_func_t)();
-
-/**
- * Return time in nanoseconds for machine. Is not wall clock time.
- *
- * @return nanoseconds since epoch for machine.
- */
-int64_t aeron_nano_clock();
-
-/**
- * Return time in milliseconds since epoch. Is wall clock time.
- *
- * @return milliseconds since epoch.
- */
-int64_t aeron_epoch_clock();
-
-/**
  * Function to return logging information.
  */
 typedef void (*aeron_log_func_t)(const char *);
