+++ aeron-driver/src/main/c/concurrent/aeron_distinct_error_log.h
@@ -21,7 +21,7 @@
 #include <stddef.h>
 #include <stdbool.h>
 #include <aeron_alloc.h>
-#include "aeronmd.h"
+#include "util/aeron_clock.h"
 #include "util/aeron_bitutil.h"
 #include "concurrent/aeron_thread.h"
 #include "concurrent/aeron_atomic.h"
